Now We can register content to our list by accessing register web page.
Also we can see the list of the registered content.

A registered content has its own  web page where one can choose
'modify', 'remove', 'go back to the list page' by buttons.

On register page, there are function that check if the content is null
or not null so that there wouldn't be a null content list that one
cannot get in.

Now we are about to make a modification function.